GGDUtils.jl

GGDUtils.jl

Installation

Using make:

make r

make u

make clean

Using Julia REPL and installing using Github url

Interpolations

Subset Tools

Types

Plot recipes

Installation

Using make:

After cloning this repo, check the make menu:

GGDUtils.jl % make help
Help Menu

make env_with_cloned_repo (or make r): Creates a Julia environment with the cloned repositories
make env_with_git_url (or make u): Creates a Julia environment with the git urls without creating local clones
make clean: Deletes Project.toml and Manifest.toml for a fresh start

make r

This option creates local copies of required private repositories at the same level as current repository and uses them in develop mode to create a Manifest.toml

make u

This option uses url of required private repositories to create a static Manifest.toml attached to current master branches of these repositories.

make clean

Deletes Manifest.toml so that environment can be recreated, to update or change the last used method.

Using Julia REPL and installing using Github url

Or, in julia REPL:

julia> using Pkg;
julia> Pkg.add(; url="https://github.com/ProjectTorreyPines/IMASDD.jl.git");
julia> Pkg.add(; url="https://github.com/ProjectTorreyPines/GGDUtils.jl.git");
julia> Pkg.instantiate()

Interpolations

Several interpolation functions are available to create interpolaiton functions for data present in a GGD represented over a particular grid subset:

GGDUtils.interp — Function

interp(
prop_values::Vector{T},
kdtree::KDTree;
use_nearest_n::Int=4,
weighing::Function=(d) -> 1 / d,

) where {T <: Real}

Lowest level interpolation function. It takes a vector of property values and a KDTree defined over a 2D space with the same number of nodes as the property values. It returns a function that can be used to interpolate the property values at any point in the space.

interp(
y::Vector{T},
TPS_mats::Tuple{Matrix{U}, Matrix{U}, Matrix{U}, Vector{Tuple{U, U}}},

) where {T <: Real, U <: Real}

Lowest level function for Thin Plate Spline method

interp(y::Vector{T}, x::Vector{Tuple{U, U}}) where {T <: Real, U <: Real}

Thin plate smoothing interpolation function for a 2d space scalar function. The algorithm has been adopted from:

http://www.geometrictools.com/Documentation/ThinPlateSplines.pdf

This is an implementation of Euler-Lagrange equation for minimizing bending energy of a surface.

interp(
prop_values::Vector{T},
space::IMASDD.edge_profiles__grid_ggd___space

) where {T <: Real}

If the whole space is provided instead of a kdtree, calculate the kdtree for whole space. Again, here it is assumed that the property values are porvided for each node of the space.

interp(
prop_values::Vector{Real},
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset

)

If a subset of the space is provided, calculate the kdtree for the subset. In this case it is assumed that the property values are provided for each element of the subset.

interp(
prop::edge_profiles__prop_on_subset,
grid_ggd::IMASDD.edge_profiles__grid_ggd,
value_field::Symbol=:values

)

Example: gridggd = dd.edgeprofiles.gridggd[1] getelectrondensity = interp(dd.edgeprofiles.ggd[1].electrons.density[1], gridggd) getefieldpar = interp(dd.edgeprofiles.ggd[1].efield[1], grid_ggd, :parallel)

interp(
prop_arr::AbstractVector{T},
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset,
value_field::Symbol=:values

) where {T <: edgeprofiles__propon_subset}

Example: sol = getgridsubsetwithindex(dd.edgeprofiles.gridggd[1], 23) getelectrondensity = interp(dd.edge_profiles.ggd[1].electrons.density, space, sol)

interp(
prop_arr::AbstractVector{T},
grid_ggd::IMASDD.edge_profiles__grid_ggd,
grid_subset_index::Int,
value_field::Symbol=:values

) where {T <: edgeprofiles__propon_subset}

Example: getnesep = interp(dd.edgeprofiles.ggd[1].electrons.density, grid_ggd, 16)

interp(eqt::IMASDD.equilibrium__time_slice)

For a given equilibrium time slice, return a function that can be used to interpolate from (r, z) space to rho (normalized toroidal flux coordinate)space.

Example: rz2rho = interp(dd.equilibrium.time_slice[1]) rho = rz2rho.([(r, z) for r in 3:0.01:9, for z in -5:0.01:5])

interp(
prop::Vector{T},
prof::IMASDD.core_profiles__profiles_1d,

) where {T <: Real}

Returns an inteprolation function for the core profile property values defined on normalized toroidal flux coordinate rho.

Example: coreprofilene = dd.coreprofiles.profiles1d[1].electrons.density getne = interp(coreprofilene, dd.coreprofiles.profiles1d[1]) getne(1) # Returns electron density at rho = 1 (separatix)

interp(
prop::Vector{T},
prof::IMASDD.core_profiles__profiles_1d,
rz2rho::Function,

)

Returns an inteprolation function in (R, Z) domain for the core profile property values defined on normalized toroidal flux coordinate rho and with a provided function to convert (R,Z) to rho.

Example:

rz2rho = interp(dd.equilibrium.timeslice[1]) coreprofilene = dd.coreprofiles.profiles1d[1].electrons.density getne = interp(coreprofilene, dd.coreprofiles.profiles1d[1], rz2rho) getn_e(5.0, 3.5) # Returns electron density at (R, Z) = (5.0, 3.5)

interp(
prop::Vector{T},
prof::IMASDD.core_profiles__profiles_1d,
eqt::IMASDD.equilibrium__time_slice,

) where {T <: Real}

Returns an inteprolation function in (R, Z) domain for the core profile property values defined on normalized toroidal flux coordinate rho and with a provided equilibrium time slice to get (R, Z) to rho conversion.

Example:

eqt = dd.equilibrium.timeslice[1] coreprofilene = dd.coreprofiles.profiles1d[1].electrons.density getne = interp(coreprofilene, dd.coreprofiles.profiles1d[1], eqt) getn_e(5.0, 3.5) # Returns electron density at (R, Z) = (5.0, 3.5)

GGDUtils.get_TPS_mats — Function

get_TPS_mats(x::Vector{Tuple{U, U}}) where {U <: Real}

We can separate matrix operations of thin plate spline method to utilize the calculation for itnerpolating over same space. Similar to the idea of reusing KDTree like above.

GGDUtils.get_kdtree — Function

get_kdtree(space::IMASDD.edge_profiles__grid_ggd___space)

Get a KDTree for all the cells in the space for search for nearest neighbours.

get_kdtree(
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset,

)

Get a KDTree for a subset of the space for search for nearest neighbours.

Subset Tools

GGDUtils.add_subset_element! — Function

add_subset_element!(
subset::IMASDD.edge_profiles__grid_ggd___grid_subset,
sn::Int,
dim::Int,
index::Int,
in_subset=(x...) -> true;
kwargs...,

)

Adds a new element to girdsubset with properties space number (sn), dimension (dim), and index (index). The element is added only if the function insubset returns true.

add_subset_element!(
subset,
sn,
dim,
index::Vector{Int},
in_subset=(x...) -> true;
kwargs...,

)

Overloaded to work differently (faster) with list of indices to be added.

GGDUtils.get_subset_space — Function

get_subset_space(space::IMASDD.edge_profiles__grid_ggd___space,
elements::AbstractVector{<:IMASDD.edge_profiles__grid_ggd___grid_subset___element})

Returns an array of space object indices corresponding to the correct objectsperdimension (nodes, edges or cells) for the subset elements.

GGDUtils.get_grid_subset — Function

get_grid_subset(
grid_ggd::IMASDD.edge_profiles__grid_ggd,
grid_subset_index::Int,

)

Returns the gridsubset in a gridggd with the matching gridsubsetindex

get_grid_subset(
grid_ggd::IMASDD.edge_profiles__grid_ggd,
grid_subset_name::String,

)

Returns the gridsubset in a gridggd with the matching gridsubsetname

GGDUtils.get_subset_boundary_inds — Function

get_subset_boundary_inds(
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset,

)

Returns an array of space object indices corresponding to the boundary of the subset. That means, it returns indices of nodes that are at the end of open edge subset or it returns the indices of edges that are the the boundary of a cell subset. Returns an empty array if the subset is 1D (nodes).

GGDUtils.get_subset_boundary — Function

get_subset_boundary(
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset,

)

Returns an array of elements of grid_subset generated from the boundary of the subset provided. The dimension of these elments is reduced by 1.

GGDUtils.subset_do — Function

subset_do(set_operator,
itrs::Vararg{AbstractVector{<:IMASDD.edge_profiles__grid_ggd___grid_subset___element}};
space::IMASDD.edge_profiles__grid_ggd___space=IMASDD.edge_profiles__grid_ggd___space(),
use_nodes=false)

Function to perform any set operation (intersect, union, setdiff etc.) on subset.element to generate a list of elements to go to subset object. If usenodes is true, the set operation will be applied on the set of nodes from subset.element, space argument is required for this. Note: that the arguments are subset.element (not the subset itself). Similarly, the return object is a list of IMASDD.edgeprofiles__gridggd___gridsubset___element.

GGDUtils.get_subset_centers — Function

get_subset_centers(space::IMASDD.edge_profiles__grid_ggd___space,
                        subset::IMASDD.edge_profiles__grid_ggd___grid_subset)

Returns an array of tuples corresponding to (r,z) coordinates of the center of cells or the center of edges in the subset space.

GGDUtils.project_prop_on_subset! — Function

project_prop_on_subset!(prop_arr::AbstractVector{T},
from_subset::IMASDD.edge_profiles__grid_ggd___grid_subset,
to_subset::IMASDD.edge_profiles__grid_ggd___grid_subset,
value_field::Symbol=:values,

) where {T <: edgeprofiles__propon_subset}

If the dimensions of fromsubset and tosubset are the same, this function can be used to add another instance on a property vector representing the value in tosubset without any interpolation or use of space object. The function returns a tuple of indices of elements of tosubset and the values of the property in to_subset.

GGDUtils.deepcopy_subset — Function

deepcopy_subset(subset::IMASDD.edge_profiles__grid_ggd___grid_subset)

Faster deepcopy function for gridsubset object. This function is used to create a deep copy of a gridsubset object bypassing several checks performed by IMASDD.

Base.:∈ — Function

Base.:∈(
point::Tuple{Real, Real},
subset_of_space::Tuple{
    IMASDD.edge_profiles__grid_ggd___grid_subset,
    IMASDD.edge_profiles__grid_ggd___space,
},

)

Overloading ∈ operator to check if a point is inside a subset of space.

If the subset is 1-dimensional, all points are searched. If the subset is 2-dimensional, it is checked if the point is within the enclosed area. It is assumed that a 2-dimensional subset used in such a context will form a closed area. If the subset is 3-dimensional, its boundary is calculated on the fly. If used multiple times, it is recommended to calculate the boundary once and store it in a variable.

GGDUtils.get_prop_with_grid_subset_index — Function

get_prop_with_grid_subset_index(
prop::AbstractVector{T},
grid_subset_index::Int,

) where {T <: edgeprofiles__propon_subset}

Find the edgeprofiles property instance in an array of properties that corresponds to the gridsubset_index provided.

Types

GGDUtils.get_types_with — Function

get_types_with(parent::Type, field::Symbol)

A type creation utility meant for searching types in IMAS database. This function returns a list of types that are fields at any level below the parent data type which have a particular field present in it.

Example:

gettypeswith(IMASDD.edgeprofiles, :gridsubset_index)

returns all edgeprofiles types that have a subfield named gridsubset_index.

Plot recipes

Several plot recipes have been defined for easy visualization.

RecipesBase.apply_recipe — Method

plot(space::IMASDD.edge_profiles__grid_ggd___space)

Plot the grid_ggd space object. Defaults to size of [600, 900] and linecolor of :black, linewidth of 0.2, and no legend.

RecipesBase.apply_recipe — Method

plot(
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset,

)

Plot the a subset of a space. Defaults to size of [600, 900] and linecolor of :black, linewidth of 0.2, and no legend.

RecipesBase.apply_recipe — Method

plot(
grid_ggd::IMASDD.edge_profiles__grid_ggd,
prop::IMASDD.IDSvectorElement,

)

Plot 2D heatmap of edgeprofilesggd property on a gridggd space object. Defaults to size of [635, 900], xaxis of "R / m", yaxis of "Z / m", and no legend. If :seriescolor is not provided, :inferno color scheme is used. If :colorbartitle is not provided, the property name is used. This function creates a plot with layout [a{0.95w} b] where a is the heatmap and b is the colorbar.

RecipesBase.apply_recipe — Method

plot(
grid_ggd_arr::AbstractVector{<:IMASDD.edge_profiles__grid_ggd},
prop::IMASDD.IDSvectorElement,

)

Plot 2D heatmap of edgeprofilesggd property on a gridggd space object. Defaults to size of [635, 900], xaxis of "R / m", yaxis of "Z / m", and no legend. If :seriescolor is not provided, :inferno color scheme is used. If :colorbartitle is not provided, the property name is used. This function creates a plot with layout [a{0.95w} b] where a is the heatmap and b is the colorbar.

RecipesBase.apply_recipe — Method

plot(
ifo::IMASDD.interferometer,

)

Plot all the channels of interferometer object. Optional keywords:

:plottype: :los(default), :ne, or :neaverage. :los plots the line of sight of the channel in a 2D plot, :ne plots the integrated ne along the line of sight vs time, and :neaverage plots the average n_e vs time.

:mirror: true(default) or false.

:mirror_length: 0.5(default).

:mirror_thickness: 0.1(default).

RecipesBase.apply_recipe — Method

plot(
ifo_ch::IMASDD.interferometer__channel,

)

Plot individual channel of interferometer. Optional keywords:

:plottype: :los(default), :ne, or :neaverage. :los plots the line of sight of the channel in a 2D plot, :ne plots the integrated ne along the line of sight vs time, and :neaverage plots the average n_e vs time.

:mirror: true(default) or false.

:mirror_length: 0.5(default).

:mirror_thickness: 0.1(default).

RecipesBase.apply_recipe — Method

plot(
ifo_ch_los::IMASDD.interferometer__channel___line_of_sight,

)

Plot line of sight of a channel of interferometer.

Default plot settings:

subplot: 1

size: [600, 900]

xaxis: "R / m"

yaxis: "Z / m"

Optional keywords:

:mirror: true(default) or false.

:mirror_length: 0.5(default).

:mirror_thickness: 0.1(default).

RecipesBase.apply_recipe — Method

plot(
ifo_ch_n_e_line::IMASDD.interferometer__channel___n_e_line,

)

Plot line integrated electron density of a channel of interferometer.

Default plot settings:

subplot: 1

xaxis: "time / s"

yaxis: "Integerated n_e / m^-2"

Optional keywords:

:average: true or false(default). If true, plot the average n_e vs time.

RecipesBase.apply_recipe — Method

plot(
ifo_ch_n_e_line_average::IMASDD.interferometer__channel___n_e_line_average,

)

Plot average electron density of a channel of interferometer.

Default plot settings:

subplot: 1

xaxis: "time / s"

yaxis: "Average n_e / m^-3"