GGDUtils.jl

Installation

Clone the repo and run make command:

% GGDUtils % make help
Help Menu

make env_with_cloned_repo (or make r): Creates a Julia environment with the cloned repositories
make env_with_git_url (or make u): Creates a Julia environment with the git urls without creating local clones
make clean: Deletes Project.toml and Manifest.toml for a fresh start

Interpolations

Several interpolation functions are available to create interpolaiton functions for data present in a GGD represented over a particular grid subset:

GGDUtils.interp — Function

interp(
prop_values::Vector{T},
kdtree::KDTree;
use_nearest_n::Int=4,
weighing::Function=(d) -> 1 / d,

) where {T <: Real}

Lowest level interpolation function. It takes a vector of property values and a KDTree defined over a 2D space with the same number of nodes as the property values. It returns a function that can be used to interpolate the property values at any point in the space.

interp(
y::Vector{T},
TPS_mats::Tuple{Matrix{U}, Matrix{U}, Matrix{U}, Vector{Tuple{U, U}}},

) where {T <: Real, U <: Real}

Lowest level function for Thin Plate Spline method

interp(y::Vector{T}, x::Vector{Tuple{U, U}}) where {T <: Real, U <: Real}

Thin plate smoothing interpolation function for a 2d space scalar function. The algorithm has been adopted from:

http://www.geometrictools.com/Documentation/ThinPlateSplines.pdf

This is an implementation of Euler-Lagrange equation for minimizing bending energy of a surface.

interp(
prop_values::Vector{T},
space::IMASDD.edge_profiles__grid_ggd___space

) where {T <: Real}

If the whole space is provided instead of a kdtree, calculate the kdtree for whole space. Again, here it is assumed that the property values are porvided for each node of the space.

interp(
prop_values::Vector{Real},
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset

)

If a subset of the space is provided, calculate the kdtree for the subset. In this case it is assumed that the property values are provided for each element of the subset.

interp(
prop::edge_profiles__prop_on_subset,
grid_ggd::IMASDD.edge_profiles__grid_ggd,
value_field::Symbol=:values

)

Example: gridggd = dd.edgeprofiles.gridggd[1] getelectrondensity = interp(dd.edgeprofiles.ggd[1].electrons.density[1], gridggd) getefieldpar = interp(dd.edgeprofiles.ggd[1].efield[1], grid_ggd, :parallel)

interp(
prop_arr::AbstractVector{T},
space::IMASDD.edge_profiles__grid_ggd___space,
subset::IMASDD.edge_profiles__grid_ggd___grid_subset,
value_field::Symbol=:values

) where {T <: edgeprofiles__propon_subset}

Example: sol = getgridsubsetwithindex(dd.edgeprofiles.gridggd[1], 23) getelectrondensity = interp(dd.edge_profiles.ggd[1].electrons.density, space, sol)

interp(
prop_arr::AbstractVector{T},
grid_ggd::IMASDD.edge_profiles__grid_ggd,
grid_subset_index::Int,
value_field::Symbol=:values

) where {T <: edgeprofiles__propon_subset}

Example: getnesep = interp(dd.edgeprofiles.ggd[1].electrons.density, grid_ggd, 16)

interp(eqt::IMASDD.equilibrium__time_slice)

For a given equilibrium time slice, return a function that can be used to interpolate from (r, z) space to rho (normalized toroidal flux coordinate)space.

Example: rz2rho = interp(dd.equilibrium.time_slice[1]) rho = rz2rho.([(r, z) for r in 3:0.01:9, for z in -5:0.01:5])

interp(
prop::Vector{T},
prof::IMASDD.core_profiles__profiles_1d,

) where {T <: Real}

Returns an inteprolation function for the core profile property values defined on normalized toroidal flux coordinate rho.

Example: coreprofilene = dd.coreprofiles.profiles1d[1].electrons.density getne = interp(coreprofilene, dd.coreprofiles.profiles1d[1]) getne(1) # Returns electron density at rho = 1 (separatix)

interp(
prop::Vector{T},
prof::IMASDD.core_profiles__profiles_1d,
rz2rho::Function,

)

Returns an inteprolation function in (R, Z) domain for the core profile property values defined on normalized toroidal flux coordinate rho and with a provided function to convert (R,Z) to rho.

Example:

rz2rho = interp(dd.equilibrium.timeslice[1]) coreprofilene = dd.coreprofiles.profiles1d[1].electrons.density getne = interp(coreprofilene, dd.coreprofiles.profiles1d[1], rz2rho) getn_e(5.0, 3.5) # Returns electron density at (R, Z) = (5.0, 3.5)

interp(
prop::Vector{T},
prof::IMASDD.core_profiles__profiles_1d,
eqt::IMASDD.equilibrium__time_slice,

) where {T <: Real}

Returns an inteprolation function in (R, Z) domain for the core profile property values defined on normalized toroidal flux coordinate rho and with a provided equilibrium time slice to get (R, Z) to rho conversion.

Example:

eqt = dd.equilibrium.timeslice[1] coreprofilene = dd.coreprofiles.profiles1d[1].electrons.density getne = interp(coreprofilene, dd.coreprofiles.profiles1d[1], eqt) getn_e(5.0, 3.5) # Returns electron density at (R, Z) = (5.0, 3.5)

GGDUtils.get_TPS_mats — Function

get_TPS_mats(x::Vector{Tuple{U, U}}) where {U <: Real}

We can separate matrix operations of thin plate spline method to utilize the calculation for itnerpolating over same space. Similar to the idea of reusing KDTree like above.